step 11, b

+++ src/main/java/com/back/domain/wiseSaying/controller/WiseSayingController.java
@@ -1,13 +1,11 @@
 package com.back.domain.wiseSaying.controller;
 
 import com.back.domain.AppContext;
+import com.back.domain.global.Rq;
 import com.back.domain.wiseSaying.entity.WiseSaying;
 import com.back.domain.wiseSaying.service.WiseSayingService;
 
-import java.util.Arrays;
-import java.util.Map;
 import java.util.Scanner;
-import java.util.stream.Collectors;
 
 public class WiseSayingController {
     private final Scanner scanner;
@@ -42,20 +40,16 @@ public class WiseSayingController {
         }
     }
 
-    public void actionDelete(String cmd) {
-        String[] cmdBits = cmd.split("\\?",2);
-        String queryString = cmdBits[1];
+    public void actionDelete(Rq rq) {
+        int id = rq.getParamAsInt("id",-1);
 
-        Map<String, String> params = Arrays
-                .stream(queryString.split("&"))
-                .map(e->e.split("=",2))
-                .filter(e->e.length == 2 & !e[0].isBlank() && !e[1].isBlank())
-                .collect(Collectors.toMap(e->e[0].trim(),e->e[1].trim()));
-
-        String idStr = params.getOrDefault("id","");
-        int id = Integer.parseInt(idStr);
+        if(id == -1 ){
+            System.out.printf("id를 정확히 입력해주세요");
+            return;
+        }
 
         boolean deleted = wiseSayingService.delete(id);
+
 
         if(deleted == false ){
             System.out.printf("%d번 명언은 존재하지 않습니다.\n", id);
@@ -66,18 +60,13 @@ public class WiseSayingController {
         System.out.printf("%d번 명언이 삭제되었습니다. \n", id);
     }
 
-    public void actionModify(String cmd) {
-        String[] cmdBits = cmd.split("\\?",2);
-        String queryString = cmdBits[1];
+    public void actionModify(Rq rq) {
+        int id = rq.getParamAsInt("id",-1);
 
-        Map<String, String> params = Arrays
-                .stream(queryString.split("&"))
-                .map(e->e.split("=",2))
-                .filter(e->e.length == 2 & !e[0].isBlank() && !e[1].isBlank())
-                .collect(Collectors.toMap(e->e[0].trim(),e->e[1].trim()));
-
-        String idStr = params.getOrDefault("id","");
-        int id = Integer.parseInt(idStr);
+        if(id == -1 ){
+            System.out.printf("id를 정확히 입력해주세요");
+            return;
+        }
 
         WiseSaying wiseSaying = wiseSayingService.findById(id);
 
@@ -86,11 +75,11 @@ public class WiseSayingController {
             return;
         }
 
-        System.out.printf("명언(기존) : $s\n", wiseSaying.getContent());
+        System.out.printf("명언(기존) : %s \n", wiseSaying.getContent());
         System.out.print("명언 : ");
         String content = scanner.nextLine();
 
-        System.out.printf("작가(기존) : $s\n", wiseSaying.getAuthor());
+        System.out.printf("작가(기존) : %s\n", wiseSaying.getAuthor());
         System.out.print("명언 : ");
         String author = scanner.nextLine();
